+++ src/Compilers/Core/CommandLine/CommandLine.projitems
@@ -10,7 +10,6 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)BuildProtocol.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)BuildClient.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ConsoleUtil.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NativeMethods.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)CompilerServerLogger.cs" />
